Bump LLVM to 7ce613fc77af092dd6e9db71ce3747b75bc5616e
