+++ lib/Bindings/Python/CMakeLists.txt
@@ -71,7 +71,6 @@ declare_mlir_dialect_python_bindings(
   TD_FILE dialects/CombOps.td
   SOURCES
     dialects/comb.py
-    dialects/_comb_ops_ext.py
   DIALECT_NAME comb)
 
 declare_mlir_dialect_python_bindings(
@@ -80,7 +79,6 @@ declare_mlir_dialect_python_bindings(
   TD_FILE dialects/ESIOps.td
   SOURCES
     dialects/esi.py
-    dialects/_esi_ops_ext.py
   DIALECT_NAME esi)
 
 declare_mlir_dialect_python_bindings(
@@ -89,7 +87,6 @@ declare_mlir_dialect_python_bindings(
   TD_FILE dialects/HWOps.td
   SOURCES
     dialects/hw.py
-    dialects/_hw_ops_ext.py
   DIALECT_NAME hw)
 
 declare_mlir_dialect_python_bindings(
@@ -98,7 +95,6 @@ declare_mlir_dialect_python_bindings(
   TD_FILE dialects/MSFTOps.td
   SOURCES
     dialects/msft.py
-    dialects/_msft_ops_ext.py
   DIALECT_NAME msft)
 
 declare_mlir_dialect_python_bindings(
@@ -115,7 +111,6 @@ declare_mlir_dialect_python_bindings(
   TD_FILE dialects/SeqOps.td
   SOURCES
     dialects/seq.py
-    dialects/_seq_ops_ext.py
   DIALECT_NAME seq)
 
 declare_mlir_dialect_python_bindings(
@@ -124,7 +119,6 @@ declare_mlir_dialect_python_bindings(
   TD_FILE dialects/SVOps.td
   SOURCES
     dialects/sv.py
-    dialects/_sv_ops_ext.py
   DIALECT_NAME sv)
 
 declare_mlir_dialect_python_bindings(
@@ -133,7 +127,6 @@ declare_mlir_dialect_python_bindings(
   TD_FILE dialects/FSMOps.td
   SOURCES
     dialects/fsm.py
-    dialects/_fsm_ops_ext.py
   DIALECT_NAME fsm)
 
 declare_mlir_dialect_python_bindings(
@@ -142,7 +135,6 @@ declare_mlir_dialect_python_bindings(
   TD_FILE dialects/HWArithOps.td
   SOURCES
     dialects/hwarith.py
-    dialects/_hwarith_ops_ext.py
   DIALECT_NAME hwarith)
 
 declare_mlir_dialect_python_bindings(
