+++ lib/Bindings/Python/dialects/comb.py
@@ -2,11 +2,12 @@
 #  See https://llvm.org/LICENSE.txt for license information.
 #  SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
 
-from ._comb_ops_gen import *
-
-from ..support import NamedValueOpView
-
+from . import comb
+from ..dialects._ods_common import _cext as _ods_cext
 from ..ir import IntegerAttr, IntegerType, OpView
+from ..support import NamedValueOpView, get_value
+from ._comb_ops_gen import *
+from ._comb_ops_gen import _Dialect
 
 
 # Sugar classes for the various possible verions of ICmpOp.
@@ -95,4 +96,196 @@ class GtUOp:
 
 @CompareOp(9)
 class GeUOp:
+  pass
+
+
+# Builder base classes for non-variadic unary and binary ops.
+class UnaryOpBuilder(NamedValueOpView):
+
+  def operand_names(self):
+    return ["input"]
+
+  def result_names(self):
+    return ["result"]
+
+
+def UnaryOp(base):
+
+  class _Class(base):
+
+    @classmethod
+    def create(cls, input=None, result_type=None):
+      mapping = {"input": input} if input else {}
+      return UnaryOpBuilder(cls, result_type, mapping)
+
+  return _Class
+
+
+class ExtractOpBuilder(UnaryOpBuilder):
+
+  def __init__(self, low_bit, data_type, input_port_mapping={}, **kwargs):
+    low_bit = IntegerAttr.get(IntegerType.get_signless(32), low_bit)
+    super().__init__(comb.ExtractOp, data_type, input_port_mapping, [],
+                     [low_bit], **kwargs)
+
+
+class BinaryOpBuilder(NamedValueOpView):
+
+  def operand_names(self):
+    return ["lhs", "rhs"]
+
+  def result_names(self):
+    return ["result"]
+
+
+def BinaryOp(base):
+
+  class _Class(base):
+
+    @classmethod
+    def create(cls, lhs=None, rhs=None, result_type=None):
+      mapping = {}
+      if lhs:
+        mapping["lhs"] = lhs
+      if rhs:
+        mapping["rhs"] = rhs
+      return BinaryOpBuilder(cls, result_type, mapping)
+
+  return _Class
+
+
+# Base classes for the variadic ops.
+def VariadicOp(base):
+
+  class _Class(base):
+
+    @classmethod
+    def create(cls, *args):
+      return cls([get_value(a) for a in args])
+
+  return _Class
+
+
+# Base class for miscellaneous ops that can't be abstracted but should provide a
+# create method for uniformity.
+def CreatableOp(base):
+
+  class _Class(base):
+
+    @classmethod
+    def create(cls, *args, **kwargs):
+      return cls(*args, **kwargs)
+
+  return _Class
+
+
+# Sugar classes for the various non-variadic unary ops.
+@_ods_cext.register_operation(_Dialect, replace=True)
+class ExtractOp(ExtractOp):
+
+  @staticmethod
+  def create(low_bit, result_type, input=None):
+    mapping = {"input": input} if input else {}
+    return ExtractOpBuilder(low_bit,
+                            result_type,
+                            mapping,
+                            needs_result_type=True)
+
+
+@UnaryOp
+@_ods_cext.register_operation(_Dialect, replace=True)
+class ParityOp(ParityOp):
+  pass
+
+
+# Sugar classes for the various non-variadic binary ops.
+@BinaryOp
+@_ods_cext.register_operation(_Dialect, replace=True)
+class DivSOp(DivSOp):
+  pass
+
+
+@BinaryOp
+@_ods_cext.register_operation(_Dialect, replace=True)
+class DivUOp(DivUOp):
+  pass
+
+
+@BinaryOp
+@_ods_cext.register_operation(_Dialect, replace=True)
+class ModSOp(ModSOp):
+  pass
+
+
+@BinaryOp
+@_ods_cext.register_operation(_Dialect, replace=True)
+class ModUOp(ModUOp):
+  pass
+
+
+@BinaryOp
+@_ods_cext.register_operation(_Dialect, replace=True)
+class ShlOp(ShlOp):
+  pass
+
+
+@BinaryOp
+@_ods_cext.register_operation(_Dialect, replace=True)
+class ShrSOp(ShrSOp):
+  pass
+
+
+@BinaryOp
+@_ods_cext.register_operation(_Dialect, replace=True)
+class ShrUOp(ShrUOp):
+  pass
+
+
+@BinaryOp
+@_ods_cext.register_operation(_Dialect, replace=True)
+class SubOp(SubOp):
+  pass
+
+
+# Sugar classes for the variadic ops.
+@VariadicOp
+@_ods_cext.register_operation(_Dialect, replace=True)
+class AddOp(AddOp):
+  pass
+
+
+@VariadicOp
+@_ods_cext.register_operation(_Dialect, replace=True)
+class MulOp(MulOp):
+  pass
+
+
+@VariadicOp
+@_ods_cext.register_operation(_Dialect, replace=True)
+class AndOp(AndOp):
+  pass
+
+
+@VariadicOp
+@_ods_cext.register_operation(_Dialect, replace=True)
+class OrOp(OrOp):
+  pass
+
+
+@VariadicOp
+@_ods_cext.register_operation(_Dialect, replace=True)
+class XorOp(XorOp):
+  pass
+
+
+@VariadicOp
+@_ods_cext.register_operation(_Dialect, replace=True)
+class ConcatOp(ConcatOp):
+  pass
+
+
+# Sugar classes for miscellaneous ops.
+@CreatableOp
+@_ods_cext.register_operation(_Dialect, replace=True)
+class MuxOp(MuxOp):
   pass
